Support table layout edit UI when geneVariant defaults to groupsetting
